browser_use_agent/requirements.txt
# Browser-Use + LangChain Together.ai Integration
# Python >=3.11 required

browser-use>=0.7.0
langchain-together>=0.3.1
python-dotenv>=1.0.0
playwright>=1.40.0
